MAINT Clean-up utils.__init__: move tests into corresponding test files (#28842)
